calc_prelim_stats: Calculates initial statistics for a Seurat object as required for downstream analysis

Calculates the percentage of mitochondrial and ribosomal reads for each cell in the Seurat object. These statistics can be used for filtering or for feature regression in downstream normalization.

Usage

calc_prelim_stats(so, sampleID = NULL)

Arguments

so

A Seurat object imported from .h5 file generated by 10X CellRanger

sampleID

An optional character argument to include sample label (useful for downstream multi-sample analysis)

Value

A Seurat object with calculated statistics added as metadata